Latest version.

+++ Structures/Structures.OBWholeOfLife/StepResults.fs
@@ -9,12 +9,12 @@ module OBWholeOfLife =
     [<NoComparison>]
     type StepResults =
         {
-            UnsmoothedAssetShare    : double
-            SmoothedAssetShare      : double
-            GuaranteedDeathBenefit  : double
-            SurrenderBenefit        : double
-            DeathBenefit            : double
-            ExitBonusRate           : double
-            UnpaidPremiums          : double
-            DeathUpliftFactor       : double
+            UnsmoothedAssetShare    : float32
+            SmoothedAssetShare      : float32
+            GuaranteedDeathBenefit  : float32
+            SurrenderBenefit        : float32
+            DeathBenefit            : float32
+            ExitBonusRate           : float32
+            UnpaidPremiums          : float32
+            DeathUpliftFactor       : float32
         }
